// src/app/claim-processing/manual-claim-processing/process-claim-drugeob/model/drug-bill-lines.model.ts
/**
 * Model class DrugBillLines
 * Path: screenbean/procclmdrugeobservice
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmdrugeobservice::DrugBillLines
 * Legacy Mapping: DRUG-BILL-LINES
 */
export class DrugBillLines {
  plnCodeA = '';
  plnCode = '';
  tosA = '';
  tos = '';
  prvNameA = '';
  prvName = '';
  dosFrmA = '';
  dosFrm = 0;
  dosToA = '';
  dosTo = 0;
  chrgAmtA = '';
  chrgAmt = '';
  chrgAmtN = 0;
  cvrdExpA = '';
  cvrdExp = '';
  cvrdExpN = 0;
  dedAmtA = '';
  dedAmt = '';
  benAmtA = '';
  benAmt = '';
  ndcA = '';
  ndc = '';
  assgnIndA = '';
  assgnInd = '';
  validateFromDate = '';
  validateToDate = '';
  serialNumber?: number;
  dateOfService?: string;
}