// src/app/claim-processing/manual-claim-processing/process-claim-drugeob/model/proc-clm-drug-eob.model.ts
import {DrugBillLines} from './drug-bill-lines.model';

/**
 * Model class ProcClmDrugEob
 * Path: screenbean/procclmdrugeobservice
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmdrugeobservice::ProcClmDrugEob
 * Legacy Mapping: DRUG-EOB-MAP
 */
export class ProcClmDrugEob {
  wholeNmA = '';
  wholeNm = '';
  acctNumA = '';
  acctNum = '';
  drugBillLines: DrugBillLines[] = [];
  assgnMaxAmtA = '';
  assgnMaxAmt = '';
  assgnMaxAmtN = 0;
  buwAmtA = '';
  buwAmt = '';
  buwAmtN = 0;
  subtotAmtA = '';
  subtotAmt = '';
  subtotAmtN = 0;
  assAdjA = '';
  assAdj = '';
  assAdjN = 0;
  adjClmA = '';
  adjClm = '';
  adjAmtA = '';
  adjAmt = '';
  adjAmtN = 0;
  assBenA = '';
  assBen = '';
  assBenN = 0;
  totBenA = '';
  totBen = '';
  totBenN = 0;
  errLineA = '';
  errLine = '';
  cmndI = '';
  mem1 = '';
  mem2I = '';
  mem3I = '';
  cno1I = '';
  cno2I = '';
  cno3I = '';
  cno4I = '';
  cno5I = '';
  cno6I = '';
  com1I = '';
  com2I = '';
  com2a = '';
  com1a = '';
  cno6a = '';
  cno5a = '';
  cno4a = '';
  cno3a = '';
  cno2a = '';
  cno1a = '';
  mem3a = '';
  mem2a = '';
  mem1a = '';
  cmnda = '';
}